Title: 씨 오 더블유, Time: 3013ms, Memory: 12MB, Status: Time Limit Exceed - CodeTree

+++ 240819/씨 오 더블유/c-o-w.java
@@ -4,19 +4,19 @@ import java.io.*;
 public class Main {
     static int n;
     static String str;
-    static int[] dp;
+    static long[] dp;
     public static void main(String[] args) throws Exception{
         BufferedReader bf= new BufferedReader(new InputStreamReader(System.in));
         n = Integer.parseInt(bf.readLine());
         str = bf.readLine();
-        dp = new int[n];
+        dp = new long[n];
         Arrays.fill(dp, -1);
         for(int i = 0;i<n;i++){
             if(str.charAt(i) == 'C') {
                 dfs(i);
             }
         }
-        int ans = 0;
+        long ans = 0;
         for(int i = 0;i<n;i++){
             if(str.charAt(i) == 'C'){
                 ans += dp[i];
@@ -26,7 +26,7 @@ public class Main {
         
     }
 
-    static int dfs(int idx){
+    static long dfs(int idx){
         
         if(str.charAt(idx) == 'W'){
             return 1;
